rjson refactoring, porting to llvm 10.0

+++ cc/color-gradient.cc
@@ -48,7 +48,7 @@ Color acmacs::color::perceptually_uniform_heatmap(size_t total_colors, size_t co
     if (color_index >= total_colors)
         color_index = total_colors - 1;
     const auto step = static_cast<double>(sScaleColorPerceptualViridis.size()) / static_cast<double>(total_colors - 1);
-    const auto offset = static_cast<size_t>(std::lround(color_index * step));
+    const auto offset = static_cast<size_t>(std::lround(static_cast<double>(color_index) * step));
     if (offset >= sScaleColorPerceptualViridis.size())
         return sScaleColorPerceptualViridis.back();
     else
@@ -86,7 +86,7 @@ std::vector<Color> acmacs::color::bezier_gradient(Color c1, Color c2, Color c3, 
 {
     const std::array colors{c1, c2, c3};
     return ranges::views::ints(0UL, output_size)
-            | ranges::views::transform([&colors,output_size](auto index) { return bezier_interpolation(colors, static_cast<double>(index) / (output_size - 1)); })
+            | ranges::views::transform([&colors,output_size](auto index) { return bezier_interpolation(colors, static_cast<double>(index) / static_cast<double>(output_size - 1)); })
             | ranges::to<std::vector>;
 
 } // acmacs::color::bezier_gradient
